requirements.txt
omegaconf
matplotlib
numpy
torch
torchvision
torchmetrics
transformers
lightning
clearml
datasets
python-dotenv
tensorboard